+++ src/main/java/com/hororok/monta/dto/response/palette/PostPaletteResponseDto.java
@@ -18,10 +18,10 @@ public class PostPaletteResponseDto {
     @AllArgsConstructor
     public static class Data{
         @JsonProperty("palette_id")
-        private Long paletteId;
+        private int paletteId;
     }
 
-    public PostPaletteResponseDto(Long paletteId) {
+    public PostPaletteResponseDto(int paletteId) {
         this.data = new Data(paletteId);
         this.status = "success";
     }
